movement script changes and a makeshift park
renamed camerafollow to follow because its function changed.
I used the other projects buildings to play around and make a park(sorry
for alot of unused files in the commit) we can just delete them all when
we get better assets i assume.
changed waypointfollow to disable follow to give back control to the
player.