added the test levels themselves
put the elevator manager into the gamemaster.  the floorbuilder does
nothing atm because the arrays are predone. we could make an auto level
builder later but right now it is useless.
the go to next scene part works also.
the test next scene is just me fucking around with the objects so its
got random objects in it